Update Avalonia
Rename ReactiveUI.Avalonia.* to Splat.Avalonia.*
Remove ReactiveUi.Avalonia
Rename Avalonia Tests
