+++ src/ReactiveUI.Avalonia.Microsoft2.Tests/Splat.Avalonia.Microsoft2.Tests.csproj
@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
+    <TargetFrameworks>net7.0</TargetFrameworks>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 
     <NoWarn>$(NoWarn);1591;CA1707;SA1600;SA1601;SA1633;CA2000</NoWarn>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
-    <DefineConstants>$(DefineConstants);MICROSOFT1</DefineConstants>
+    <DefineConstants>$(DefineConstants);MICROSOFT2</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
@@ -37,23 +37,20 @@
     <Compile Include="..\ReactiveUI.Avalonia.Tests\Mocks\Views\FooView.axaml.cs" Link="Mocks\Views\FooView.axaml.cs" />
   </ItemGroup>
 
+
   <ItemGroup>
-    <ProjectReference Include="..\ReactiveUI.Avalonia.Microsoft.Extensions.DependencyInjection\ReactiveUI.Avalonia.Microsoft.Extensions.DependencyInjection.csproj" />
+    <Compile Include="..\ReactiveUI.Avalonia.Microsoft1.Tests\AvaloniaUIThreadTestsMicrosoft.cs" Link="AvaloniaUIThreadTestsMicrosoft.cs" />
   </ItemGroup>
 
   <ItemGroup>
-    <!--
-    <ProjectReference Include="..\ReactiveUI.Avalonia.Ninject\ReactiveUI.Avalonia.Ninject.csproj" />
-    -->
+    <ProjectReference Include="..\ReactiveUI.Avalonia.Microsoft.Extensions.DependencyInjection\Splat.Avalonia.Microsoft.Extensions.DependencyInjection.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Avalonia.Desktop" Version="11.0.4" />
     <PackageReference Include="Avalonia.Headless.XUnit" Version="11.0.4" />
     <PackageReference Include="Avalonia.Themes.Fluent" Version="11.0.4" />
     <PackageReference Include="Avalonia.Fonts.Inter" Version="11.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Mocks\ViewModels\" />
-    <Folder Include="Mocks\Views\" />
   </ItemGroup>
 
 </Project>
