+++ src/ReactiveUI.Avalonia.Ninject/Splat.Avalonia.Ninject.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;net6.0;net7.0</TargetFrameworks>
@@ -7,16 +7,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\ReactiveUI.Avalonia\*.cs" LinkBase="Avalonia" />
-  </ItemGroup>
-
-  <ItemGroup>
     <Compile Include="..\Splat.Ninject\NinjectDependencyResolver.cs" Link="NinjectDependencyResolver.cs" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Avalonia" Version="11.0.4" />
-    <PackageReference Include="ReactiveUI" Version="19.4.1" />
+    <PackageReference Include="Avalonia.ReactiveUI" Version="11.0.4" />
     <PackageReference Include="Ninject" Version="3.3.6" />
     <ProjectReference Include="..\Splat\Splat.csproj" />
   </ItemGroup>
