--- src/ReactiveUI.Avalonia/Attributes.cs
@@ -1,8 +1,0 @@
-// Copyright (c) 2021 .NET Foundation and Contributors. All rights reserved.
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for full license information.
-
-using Avalonia.Metadata;
-
-[assembly: XmlnsDefinition("http://reactiveui.net", "Avalonia.ReactiveUI")]
